Wysiwyg editor as component
