+++ bot/admin/web/src/app/shared/components/index.ts
@@ -36,3 +36,4 @@ export * from './sentences-generation/sentences-generation.component';
 export * from './info-button/info-button.component';
 export * from './select-bot/select-bot.component';
 export * from './date-range/date-range-calendar.component';
+export * from './wysiwyg-editor/wysiwyg-editor.component';
