+++ bot/admin/web/src/app/shared/components/wysiwyg-editor/wysiwyg-editor.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { WysiwygEditorComponent } from './wysiwyg-editor.component';
+
+describe('WysiwygEditorComponent', () => {
+  let component: WysiwygEditorComponent;
+  let fixture: ComponentFixture<WysiwygEditorComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [WysiwygEditorComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(WysiwygEditorComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
