+++ bot/admin/web/src/app/theme/styles/styles.scss
@@ -64,4 +64,5 @@ $table-color: inherit;
   // @include nb-overrides();
 }
 
+@import './jodit';
 @import './utilities';
