+++ bot/admin/web/src/app/theme/styles/utilities.scss
@@ -228,18 +228,3 @@ nb-menu {
   border-color: var(--input-basic-disabled-border-color);
   color: var(--input-basic-disabled-text-color);
 }
-
-.jodit_theme_default {
-  --jd-border-radius-default: var(--input-rectangle-border-radius);
-  --jd-color-background-default: var(--input-basic-background-color);
-  --jd-color-border: var(--input-basic-border-color);
-}
-
-.nb-theme-dark {
-  --jd-dark-toolbar-color: var(--input-basic-background-color);
-  .jodit_theme_dark {
-    --jd-dark-background-color: var(--input-basic-background-color);
-    --jd-color-background-default: var(--input-basic-background-color);
-    --jd-color-border: var(--input-basic-border-color);
-  }
-}
